pyproject.toml
[project]
name = "docmind"
version = "0.1.0"
description = "基于 LangChain + GLM 的智能文档摘要服务，支持 API 和 MCP 协议"
authors = [{name = "hrronaldo", email = "hrronaldo@163.com"}]
requires-python = "==3.12.7"
dependencies = [
    "fastapi==0.135.1",
    "uvicorn[standard]==0.41.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "langchain==1.2.10",
    "langchain-core==1.2.17",
    "langchain-community==0.4.1",
    "zhipuai==2.1.5.20250825",
    "trafilatura==2.0.0",
    "requests==2.32.5",
    "beautifulsoup4==4.14.3",
    "lxml==6.0.2",
    "python-multipart==0.0.22",
    "httpx==0.28.1",
    "fastmcp==2.13.3",
    "mcp==1.19.0",
    "sniffio==1.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []
